Plane scallar

+++ src/utils/math.c
@@ -6,7 +6,7 @@
 /*   By: dda-cunh <dda-cunh@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/03 16:43:28 by dda-cunh          #+#    #+#             */
-/*   Updated: 2024/01/18 15:04:01 by dda-cunh         ###   ########.fr       */
+/*   Updated: 2024/01/22 15:52:55 by dda-cunh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,12 +22,12 @@ t_vec3	vec3_cross_product(t_vec3 a, t_vec3 b)
 	};
 }
 
-double	vec3_dot_product(t_vec3 a, t_vec3 b)
+float	vec3_dot_product(t_vec3 a, t_vec3 b)
 {
 	return (a.x * b.x + a.y * b.y + a.z * b.z);
 }
 
-double	point3_distance_point3(t_point3 p1, t_point3 p2)
+float	point3_distance_point3(t_point3 p1, t_point3 p2)
 {
 	return (sqrt(pow((p2.x - p1.x), 2) + pow((p2.y - p1.y), 2)
 			+ pow((p2.z - p1.z), 2)));
@@ -53,7 +53,7 @@ t_point3	point3_plus_vec3(t_point3 point, t_vec3 vector)
 		});
 }
 
-t_vec3	scale_vec3(t_vec3 vector, double scalar)
+t_vec3	scale_vec3(t_vec3 vector, float scalar)
 {
 	return ((t_vec3)
 		{
@@ -65,7 +65,7 @@ t_vec3	scale_vec3(t_vec3 vector, double scalar)
 
 t_vec3	normalize_vec3(t_vec3 vector)
 {
-	double	magnitude;
+	float	magnitude;
 
 	magnitude = sqrt(vector.x * vector.x + vector.y * vector.y
 			+ vector.z * vector.z);
